test: 修复测试
